+++ src/test/java/info/macias/kaconf/test/SomeSubclass.java
@@ -2,9 +2,6 @@ package info.macias.kaconf.test;
 
 import info.macias.kaconf.Property;
 
-/**
- * Created by mmacias on 20/11/16.
- */
 public class SomeSubclass extends SomeClass {
     @Property("subpublicint")
     public int subpublicint;