+++ src/test/java/info/macias/kaconf/test/WithBooleans.java
@@ -2,9 +2,6 @@ package info.macias.kaconf.test;
 
 import info.macias.kaconf.Property;
 
-/**
- * Created by mmacias on 24/11/16.
- */
 public class WithBooleans {
     @Property("prop.true")
     public boolean trueProperty;