+++ src/test/java/info/macias/kaconf/test/WithStaticValues.java
@@ -2,9 +2,6 @@ package info.macias.kaconf.test;
 
 import info.macias.kaconf.Property;
 
-/**
- * Created by mmacias on 21/11/16.
- */
 public class WithStaticValues {
     @Property("publicint")
     public static int STATIC_VALUE;