+++ src/test/java/info/macias/kaconf/test/WithStaticValuesSubClass.java
@@ -2,9 +2,6 @@ package info.macias.kaconf.test;
 
 import info.macias.kaconf.Property;
 
-/**
- * Created by mmacias on 21/11/16.
- */
 public class WithStaticValuesSubClass extends WithStaticValues {
     @Property("subprivatebyte")
     private static byte SUB_STATIC_VALUE;