+++ src/test/kotlin/info/macias/kaconf/ConfiguratorBuilderTest.kt
@@ -3,9 +3,6 @@ package info.macias.kaconf
 import info.macias.kaconf.sources.JavaUtilPropertySource
 import junit.framework.TestCase
 
-/**
- * Created by mmacias on 16/11/16.
- */
 class ConfiguratorBuilderTest : TestCase("Test Configurator Builder") {
 
     fun testInvalidBuild() {